+++ src/main/kotlin/com/vacgom/backend/application/auth/VacgomSignupService.kt
@@ -5,18 +5,18 @@ import com.vacgom.backend.application.auth.dto.response.AuthResponse
 import com.vacgom.backend.application.auth.dto.response.MemberResponse
 import com.vacgom.backend.application.auth.dto.response.TokenResponse
 import com.vacgom.backend.domain.auth.constants.Role
+import com.vacgom.backend.domain.inoculation.Inoculation
+import com.vacgom.backend.domain.inoculation.constants.Vaccination
 import com.vacgom.backend.domain.member.HealthProfile
 import com.vacgom.backend.domain.member.MemberDetails
 import com.vacgom.backend.domain.member.Nickname
-import com.vacgom.backend.domain.vaccine.Inoculation
-import com.vacgom.backend.domain.vaccine.constants.Vaccination
 import com.vacgom.backend.exception.member.MemberError
 import com.vacgom.backend.exception.member.NicknameError
 import com.vacgom.backend.global.exception.error.BusinessException
 import com.vacgom.backend.global.security.jwt.JwtFactory
+import com.vacgom.backend.infrastructure.inoculation.persistence.InoculationRepository
 import com.vacgom.backend.infrastructure.member.persistence.HealthProfileRepository
 import com.vacgom.backend.infrastructure.member.persistence.MemberRepository
-import com.vacgom.backend.infrastructure.vaccine.persistence.InoculationRepository
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 import java.util.*
